EPMRPP-87119 || move seo constants
