--- src/pages/constants.ts
@@ -1,62 +1,0 @@
-export const SEO_DATA = {
-    index: {
-        title: 'ReportPortal test automation analytics platform and real-time reporting, powered by Machine Learning',
-        description: 'Centralized test automation dashboard. Provides AI-based defects triage and real time test report dashboard.',
-    },
-    features: {
-        title: 'Features',
-        description: 'Explore features for seamless test automation. Rethink end-to-end testing with automated defect triaging and a dynamic QA metrics dashboard.',
-    },
-    blog: {
-        title: 'Blog',
-        description: 'Stay informed by reading ReportPortal\'s updates, news, and technology articles.',
-    },
-    terms: {
-        title: 'Terms and Conditions',
-        description: 'Review the Terms and Conditions to stay informed about the rules governing the use of our test automation results dashboard.',
-    },
-    caseStudies: {
-        title: 'Case Studies',
-        description: 'Discover how organizations have optimized their testing processes thanks to ReportPortal.',
-    },
-    installation: {
-        title: 'Installation guide',
-        description: 'Three steps to get started with our test automation dashboard.',
-    },
-    qeConsulting: {
-        title: 'Quality Engineering Consulting',
-        description: 'Achieve a product-quality transformation with our outstanding Quality Engineering Consulting.',
-    },
-    testAutomationAsAService: {
-        title: 'Test Automation as a Service',
-        description: 'Enhance your testing process by utilizing our on-demand automation tools, infrastructure, hardware, and expert resources.',
-    },
-    testingAsAService: {
-        title: 'Testing as a Service',
-        description: 'Optimize your testing costs with our on-demand software, hardware, and expert resources.',
-    },
-    saas: {
-        title: 'Pricing: Saas',
-        description: 'Explore our flexible pricing plans from small teams to global enterprises. Harness the power of our end-to-end testing reporting tool.',
-    },
-    onPremises: {
-        title: 'Pricing: On-Premises',
-        description: 'Select the pricing plan that best suits your needs and unlock the full potential of our test report dashboard.',
-    },
-    qasp: {
-        title: 'Pricing for accelerators: QaSpace',
-        description: 'Dive into the specifics of our offerings to keep track of your project\'s health and progress with our advanced QA metrics dashboard.',
-    },
-    hlm: {
-        title: 'Pricing for accelerators: Healenium',
-        description: 'Choose the pricing package that enables you to stay ahead of the curve with our comprehensive test automation reporting solutions.',
-    },
-    d4j: {
-        title: 'Pricing for accelerators: Drill4J',
-        description: 'Enhance your testing strategy with our priced plans, designed to streamline your quality assurance workflows.',
-    },
-    contactUsGeneral: {
-        title: 'Contact us',
-        description: 'Whether you have questions about our services or need assistance, feel free to connect with us using this form.',
-    },
-};
