+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { Layout } from '@app/components';
 import { LandingPage } from '@app/containers/LandingPage';
-import { SEO_DATA } from "@app/pages/constants";
+import { SEO_DATA } from "@app/utils";
 
 const Root: FC = () => (
   <Layout seoData={SEO_DATA.index}>
