+++ src/pages/installation.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { Layout } from '@app/components';
 import { InstallationPage } from '@app/containers/InstallationPage';
-import { SEO_DATA } from "@app/pages/constants";
+import { SEO_DATA } from "@app/utils";
 
 const Installation: FC = () => (
   <Layout seoData={SEO_DATA.installation}>
