+++ src/pages/legal/terms.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { Layout } from '@app/components';
 import { TermsPage } from '@app/containers/TermsPage';
-import { SEO_DATA } from "@app/pages/constants";
+import { SEO_DATA } from "@app/utils";
 
 const Terms: FC = () => (
   <Layout seoData={SEO_DATA.terms}>
