+++ src/pages/testing-as-a-service.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { Layout } from '@app/components';
 import { TestingAsServicePage } from '@app/containers/TestingAsServicePage';
-import { SEO_DATA } from "@app/pages/constants";
+import { SEO_DATA } from "@app/utils";
 
 const TestingAsService: FC = () => (
   <Layout seoData={SEO_DATA.testingAsAService}>
