+++ src/templates/contact-us/contact-us.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 import { PageProps } from 'gatsby';
 import { Layout } from '@app/components';
 import { ContactUsPage } from '@app/containers/ContactUsPage';
-import { SEO_DATA } from "@app/pages/constants";
+import { SEO_DATA } from "@app/utils";
 
 interface ContextProps {
   config: {
