+++ src/templates/pricing/saas.tsx
@@ -2,8 +2,7 @@ import React, { FC } from 'react';
 import { PageProps } from 'gatsby';
 import { Layout } from '@app/components';
 import { SaasPage } from '@app/containers/SassPage';
-import { SassPricingConfig } from '@app/utils';
-import { SEO_DATA } from "@app/pages/constants";
+import { SassPricingConfig, SEO_DATA } from '@app/utils';
 
 const Saas: FC<PageProps<null, SassPricingConfig>> = ({ pageContext }) => (
   <Layout seoData={SEO_DATA.saas} className="offer-page-wrapper">
